+++ ui/src/pages/namespace/Explorer/Consumer/ConsumerEditor/Form/index.tsx
@@ -10,7 +10,7 @@ import {
 import { FC } from "react";
 import { Fieldset } from "~/components/Form/Fieldset";
 import Input from "~/design/Input";
-import { ObjArrayInput } from "~/components/Form/ObjArrayInput";
+import { ObjArrayInput } from "~/components/Form/ArrayInput";
 import { treatEmptyStringAsUndefined } from "../../../utils";
 import { useTranslation } from "react-i18next";
 import { zodResolver } from "@hookform/resolvers/zod";
