+++ ui/src/pages/namespace/Explorer/Endpoint/EndpointEditor/Form/plugins/Inbound/AclForm.tsx
@@ -8,7 +8,7 @@ import FormErrors, { errorsType } from "~/components/FormErrors";
 
 import { Fieldset } from "~/components/Form/Fieldset";
 import Input from "~/design/Input";
-import { ObjArrayInput } from "~/components/Form/ObjArrayInput";
+import { ObjArrayInput } from "~/components/Form/ArrayInput";
 import { PluginWrapper } from "../components/Modal";
 import { useTranslation } from "react-i18next";
 import { zodResolver } from "@hookform/resolvers/zod";
