+++ ui/src/pages/namespace/Explorer/Service/ServiceEditor/Form/index.tsx
@@ -23,7 +23,7 @@ import {
 import { FC } from "react";
 import { Fieldset } from "~/components/Form/Fieldset";
 import Input from "~/design/Input";
-import { ObjArrayInput } from "../../../../../../components/Form/ObjArrayInput";
+import { ObjArrayInput } from "../../../../../../components/Form/ArrayInput";
 import { useTranslation } from "react-i18next";
 import { zodResolver } from "@hookform/resolvers/zod";
